Drawing the floor after button click

+++ app/src/main/java/com/inveitix/android/clue/cmn/MapPoint.java
@@ -4,12 +4,12 @@ package com.inveitix.android.clue.cmn;
  * Created by Tito on 28.12.2015 г..
  */
 public class MapPoint {
-    public double x;
-    public double y;
+    public float x;
+    public float y;
 
     public MapPoint() {
     }
-    public MapPoint(double x, double y) {
+    public MapPoint(float x, float y) {
         this.x = x;
         this.y = y;
     }
